Fixed Custom Cost window and step calculations.
